+++ src/main/java/de/jeff_media/lumberjack/commands/CommandLumberjack.java
@@ -1,7 +1,7 @@
 package de.jeff_media.lumberjack.commands;
 
-import com.jeff_media.jefflib.BlockTracker;
 import de.jeff_media.lumberjack.LumberJack;
+import de.jeff_media.lumberjack.utils.BlockTracker;
 import org.bukkit.ChatColor;
 import org.bukkit.block.Block;
 import org.bukkit.command.Command;
@@ -10,7 +10,6 @@ import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 import org.jetbrains.annotations.NotNull;
 
-import java.util.Objects;
 
 public class CommandLumberjack implements CommandExecutor {
 
@@ -34,7 +33,7 @@ public class CommandLumberjack implements CommandExecutor {
             return true;
         }
 
-        if (args.length > 0 && args[0].equalsIgnoreCase("debug") && sender.hasPermission("lumberjack.debug")) {
+        if (args.length > 0 && args[0].equalsIgnoreCase("debug") && sender.hasPermission("lumberjack.debug") && sender instanceof Player) {
             Player player = (Player) sender;
             Block target = player.getTargetBlock(null, 20);
             player.sendMessage(String.valueOf(BlockTracker.isPlayerPlacedBlock(target)));
@@ -47,7 +46,7 @@ public class CommandLumberjack implements CommandExecutor {
         }
         Player p = (Player) sender;
         if (!sender.hasPermission("lumberjack.use")) {
-            sender.sendMessage(Objects.requireNonNull(Objects.requireNonNull(plugin.getCommand("lumberjack")).getPermissionMessage()));
+            sender.sendMessage(ChatColor.RED + "You don't have permission to use this command.");
             return true;
         }