+++ src/main/java/de/jeff_media/lumberjack/listeners/BlockBreakListener.java
@@ -1,13 +1,13 @@
 package de.jeff_media.lumberjack.listeners;
 
-import com.jeff_media.jefflib.BlockTracker;
-import com.jeff_media.jefflib.NBTAPI;
 import de.jeff_media.lumberjack.LumberJack;
 import de.jeff_media.lumberjack.NBTKeys;
 import de.jeff_media.lumberjack.NBTValues;
 import de.jeff_media.lumberjack.data.AxeMaterial;
+import de.jeff_media.lumberjack.utils.BlockTracker;
 import de.jeff_media.lumberjack.utils.TreeUtils;
 import org.bukkit.Material;
+import org.bukkit.NamespacedKey;
 import org.bukkit.block.Block;
 import org.bukkit.block.BlockFace;
 import org.bukkit.block.data.BlockData;
@@ -18,6 +18,7 @@ import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
 import org.bukkit.event.block.BlockBreakEvent;
+import org.bukkit.persistence.PersistentDataType;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -176,11 +177,11 @@ public class BlockBreakListener implements Listener {
         // falling Blocks being spawned isntead of logs
         // that were on the ground, so they broke immediately and dropped themself. I
         // think I fixed this by the following line
-        // if(logAbove.getRelative(BlockFace.DOWN).getType() == Material.AIR ||
+        // if(logAbove.getRelative(BlockFace.DOWN).getType().isAir() ||
         // logs.contains(logAbove) ||
         // logs.contains(logAbove.getRelative(BlockFace.DOWN))) {
         for (Block logAbove : logs) {
-            if (logAbove.getRelative(BlockFace.DOWN).getType() == Material.AIR || logs.contains(logAbove)
+            if (logAbove.getRelative(BlockFace.DOWN).getType().isAir() || logs.contains(logAbove)
                     || logs.contains(logAbove.getRelative(BlockFace.DOWN))) {
 
                 BlockData blockData = logAbove.getBlockData().clone();
@@ -188,7 +189,7 @@ public class BlockBreakListener implements Listener {
                 FallingBlock fallingBlock = logAbove.getLocation().getWorld()
                         .spawnFallingBlock(logAbove.getLocation().add(plugin.fallingBlockOffset), blockData);
                 if (plugin.getConfig().getBoolean("prevent-torch-exploit")) {
-                    NBTAPI.addNBT(fallingBlock, NBTKeys.IS_FALLING_LOG, NBTValues.TRUE);
+                    fallingBlock.getPersistentDataContainer().set(new NamespacedKey(plugin, NBTKeys.IS_FALLING_LOG), PersistentDataType.STRING, NBTValues.TRUE);
                 }
                 if (plugin.getConfig().getBoolean("prevent-torch-exploit-aggressive")) {
                     fallingBlock.setDropItem(false);