+++ src/main/java/de/jeff_media/lumberjack/utils/DecayUtils.java
@@ -14,20 +14,6 @@ public class DecayUtils {
 
     private static final int MAX_DISTANCE = 6;
     private static final int RADIUS = 5;
-    private static final boolean IS_AT_LEAST_v1_17;
-    private static Boolean tagsAvailable = null;
-
-    static {
-        boolean azaleaLeavesAvailable = false;
-        try {
-            //noinspection ResultOfMethodCallIgnored
-            Material.FLOWERING_AZALEA_LEAVES.getData();
-            azaleaLeavesAvailable = true;
-        } catch (Throwable ignored) {
-
-        }
-        IS_AT_LEAST_v1_17 = azaleaLeavesAvailable;
-    }
 
     public static Collection<Block> getLeaves(BlockState originalLeaf) {
         Collection<Block> blocks = new HashSet<>();
@@ -40,7 +26,6 @@ public class DecayUtils {
                 for (int z = blockZ - RADIUS; z <= blockZ + RADIUS; z++) {
                     Block candidate = world.getBlockAt(x, y, z);
                     if(candidate == originalLeaf.getBlock()) continue;
-                    if(candidate==null) continue;
                     if (candidate.getType().isAir()) continue;
                     if (!isLeaf(candidate)) {
                         continue;
@@ -48,6 +33,7 @@ public class DecayUtils {
                     if (!isMatchingLeaf(originalLeaf.getType(), candidate.getType())) {
                         continue;
                     }
+                    if (!(candidate.getBlockData() instanceof Leaves)) continue;
                     Leaves leaves = (Leaves) candidate.getBlockData();
                     if (leaves.isPersistent()) {
                         continue;
@@ -67,32 +53,12 @@ public class DecayUtils {
     }
 
     private static boolean isLeaf(Material material) {
-        if (tagsAvailable == null) {
-            try {
-                Tag.LEAVES.isTagged(material);
-                tagsAvailable = true;
-            } catch (Throwable t) {
-                tagsAvailable = false;
-            }
-        }
-
-        if (tagsAvailable) {
-            return Tag.LEAVES.isTagged(material);
-        } else {
-            return material.name().endsWith("_LEAVES");
-        }
-
+        return Tag.LEAVES.isTagged(material);
     }
 
     private static boolean isMatchingLeaf(Material leaf1, Material leaf2) {
-        if (IS_AT_LEAST_v1_17) {
-            switch (leaf1) {
-                case AZALEA_LEAVES:
-                case FLOWERING_AZALEA_LEAVES:
-                    return leaf2 == Material.AZALEA_LEAVES || leaf2 == Material.FLOWERING_AZALEA_LEAVES;
-                default:
-                    return leaf1 == leaf2;
-            }
+        if (leaf1 == Material.AZALEA_LEAVES || leaf1 == Material.FLOWERING_AZALEA_LEAVES) {
+            return leaf2 == Material.AZALEA_LEAVES || leaf2 == Material.FLOWERING_AZALEA_LEAVES;
         }
         return leaf1 == leaf2;
     }